+++ db/upgrade.php
@@ -85,5 +85,9 @@ function xmldb_local_delegateaccount_upgrade(int $oldversion): bool {
         upgrade_plugin_savepoint(true, 2026082100, 'local', 'delegateaccount');
     }
 
+    if ($oldversion < 2026082102) {
+        upgrade_plugin_savepoint(true, 2026082102, 'local', 'delegateaccount');
+    }
+
     return true;
 }